requirements.txt
appdirs==1.4.3
asgiref==3.2.7
beautifulsoup4==4.8.0
bs4==0.0.1
certifi==2019.6.16
chardet==3.0.4
click==7.1.2
cssselect==1.1.0
Django==3.0.6
djangorestframework==3.11.0
docutils==0.15.2
facebook-scraper==0.1.3
fake-useragent==0.1.11
Flask==1.1.2
Flask-SocketIO==4.3.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.11.2
Kivy==1.11.1
Kivy-Garden==0.1.4
lxml==4.4.1
MarkupSafe==1.1.1
parse==1.12.1
psycopg2==2.8.2
pyee==6.0.0
Pygments==2.4.2
pyppeteer==0.0.25
pyquery==1.4.0
python-engineio==3.13.0
python-socketio==4.6.0
pytz==2020.1
requests==2.22.0
requests-html==0.10.0
six==1.12.0
soupsieve==1.9.3
sqlparse==0.3.1
tqdm==4.35.0
urllib3==1.25.3
virtualenv==16.6.0
w3lib==1.21.0
websockets==8.0.2
Werkzeug==1.0.1
